+++ src/features/post/view-posts/index.ts
@@ -1,3 +1,4 @@
 export { usePostsBrowseParams } from "./model/usePostsBrowseParams"
+export { usePostsList } from "./model/usePostsList"
 export type { PostsBrowseParams } from "./model/usePostsBrowseParams"
 export { PostsFilterBar } from "./ui/PostsFilterBar"
