+++ src/pages/posts-manager/ui/PostsManagerPage.tsx
@@ -1,11 +1,8 @@
 import { Plus } from "lucide-react"
 import { Button, Card, CardContent, CardHeader, CardTitle } from "@shared/ui"
-import { usePostsBrowseParams, PostsFilterBar } from "@features/post/browse-posts"
-import { useQuery, keepPreviousData } from "@tanstack/react-query"
-import { postQueries } from "@entities/post/api/queries"
-import { userQueries } from "@entities/user/api/queries"
-import { Post } from "@entities/post/model"
-import { User } from "@entities/user/model"
+import { usePostsBrowseParams, usePostsList, PostsFilterBar } from "@/features/post/view-posts"
+import type { Post } from "@entities/post/model"
+import type { User } from "@entities/user/model"
 import { useUserProfileDialog } from "@features/user/view-profile/lib/useUserProfileDialog"
 import { PostsTable } from "@widgets/posts-table/ui"
 import { useAddPost } from "@features/post/add-post"
@@ -16,59 +13,11 @@ import { PostsPagination } from "@widgets/posts-pagination"
 
 const PostsManager = () => {
   const { params, setParams, stepPrev, stepNext } = usePostsBrowseParams()
-  const { skip, limit, searchQuery, sortBy, order, tag: selectedTag } = params
+  const { skip, limit, searchQuery, tag: selectedTag } = params
 
   const { openProfile, overlay: profileOverlay } = useUserProfileDialog()
 
-  const postQuery = useQuery({
-    ...postQueries.listQuery({
-      limit,
-      skip,
-      sortBy: sortBy && sortBy !== "none" ? sortBy : "id",
-      order,
-    }),
-    enabled: !searchQuery && (!selectedTag || selectedTag === "all"),
-    placeholderData: keepPreviousData,
-  })
-
-  const { data: users } = useQuery(userQueries.listQuery())
-
-  const { data: tags } = useQuery(postQueries.tagQuery())
-
-  // TODO: 이걸 나중에 묶는걸 고려해봐도...
-  const searchedPostsQuery = useQuery({
-    ...postQueries.searchQuery({
-      search: searchQuery,
-      limit,
-      skip,
-      sortBy: sortBy && sortBy !== "none" ? sortBy : "id",
-      order,
-    }),
-    enabled: !!searchQuery,
-    placeholderData: keepPreviousData,
-  })
-
-  const tagPostsQuery = useQuery({
-    ...postQueries.listByTagQuery({
-      tag: selectedTag,
-      limit,
-      skip,
-      sortBy: sortBy && sortBy !== "none" ? sortBy : "id",
-      order,
-    }),
-    enabled: !!selectedTag && selectedTag !== "all",
-    placeholderData: keepPreviousData,
-  })
-  const active = searchQuery ? searchedPostsQuery : selectedTag && selectedTag !== "all" ? tagPostsQuery : postQuery
-
-  const isLoading = active.isLoading
-
-  const posts =
-    active.data?.posts?.map((post: Post) => ({
-      ...post,
-      author: users?.users?.find((user: User) => user.id === post.userId),
-    })) ?? []
-  const total = active.data?.total ?? 0
+  const { posts, total, isLoading, tags } = usePostsList()
 
   const { deletePost } = useDeletePost()
